# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Sealed Invoices API — privacy-first invoicing anchored on the Arc testnet, with real user accounts
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices/{invoiceId}/grants:
    get:
      operationId: listGrants
      tags: [grants]
      summary: List access grants for an invoice (owner only)
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Grant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    post:
      operationId: createGrant
      tags: [grants]
      summary: Grant another user time-limited access (key wrapped in the owner's browser)
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantInput"
      responses:
        "201":
          description: Created grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grant"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Only the invoice's owner can grant access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /invoices/{invoiceId}/grants/{grantId}/revoke:
    post:
      operationId: revokeGrant
      tags: [grants]
      summary: Revoke an access grant immediately (the signed-in user must be the grantor)
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema: { type: string }
        - name: grantId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Revoked grant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Grant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Only the user who issued the grant can revoke it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  responses:
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiMessage"
  schemas:
...
    ApiMessage:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    GrantInput:
      type: object
      required: [granteeId, expiresAt, wrappedKey, granteePublicKeyJwk]
      properties:
        granteeId: { type: string }
        expiresAt: { type: string }
        wrappedKey:
          type: string
          description: AES key re-wrapped for the grantee, produced in the grantor's browser
        granteePublicKeyJwk:
          type: string
          description: >-
            Echo of the exact registered public key the wrap was prepared for.
            Verified under a row lock so a share sealed for a key the grantee
            just rotated or reset away is refused instead of stored.
    Grant:
      type: object
      required:
        [id, invoiceId, grantorId, granteeId, granteeName, expiresAt, status, createdAt]
      properties:
        id: { type: string }
        invoiceId: { type: string }
        grantorId: { type: string }
        granteeId: { type: string }
        granteeName: { type: string }
        expiresAt: { type: string }
        status:
          type: string
          enum: [active, expired, revoked]
        revokedAt: { type: ["string", "null"] }
        createdAt: { type: string }
...